requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
pymongo==4.6.1
google-generativeai==0.3.2
python-dotenv==1.0.0
faker==22.6.0
networkx==3.2.1
pyvis==0.3.2
great-expectations==0.18.8
scikit-learn==1.4.0
seaborn==0.13.2
matplotlib==3.8.2
openpyxl==3.1.2
jsonschema==4.21.0
pydantic==2.6.0
streamlit-extras==0.3.6
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5
requests==2.31.0
